frontend/requirements.txt
opencv-python
ultralytics
pillow
requests
flask
